move functions from cli into controller
